requirements.txt
pandas
selenium
webdriver-manager
SQLAlchemy
psycopg2-binary
python-dotenv